Move CommonUtils to the utils module
